Feature/dev room image api

+++ src/main/java/KL/KL_Booking_App/service/impl/RoomServiceImpl.java
@@ -1,21 +1,26 @@
 package KL.KL_Booking_App.service.impl;
 
-import KL.KL_Booking_App.entity.Hotel;
-import KL.KL_Booking_App.entity.Room;
-import KL.KL_Booking_App.entity.RoomImage;
+import KL.KL_Booking_App.entity.*;
 import KL.KL_Booking_App.entity.roomType.RoomType;
 import KL.KL_Booking_App.exeption.ResourceNotFoundException;
-import KL.KL_Booking_App.payload.response.HotelDto;
 import KL.KL_Booking_App.payload.response.RoomDto;
 import KL.KL_Booking_App.payload.response.RoomImageDto;
+import KL.KL_Booking_App.payload.response.RoomResponse;
+import KL.KL_Booking_App.repository.LocationRepository;
 import KL.KL_Booking_App.repository.RoomRepository;
 import KL.KL_Booking_App.service.IHotelService;
+import KL.KL_Booking_App.service.IReservationService;
 import KL.KL_Booking_App.service.IRoomService;
-import KL.KL_Booking_App.utils.HotelUtils;
+import KL.KL_Booking_App.utils.RoomUtils;
 import jakarta.transaction.Transactional;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.stream.Collectors;
 
 @Service
 public class RoomServiceImpl implements IRoomService {
@@ -24,9 +29,15 @@ public class RoomServiceImpl implements IRoomService {
 
     private final IHotelService hotelService;
 
-    public RoomServiceImpl(RoomRepository roomRepository, IHotelService hotelService) {
+    private final RoomUtils roomUtils;
+
+    private final LocationRepository locationRepository;
+
+    public RoomServiceImpl(RoomRepository roomRepository, IHotelService hotelService, RoomUtils roomUtils, LocationRepository locationRepository) {
         this.roomRepository = roomRepository;
         this.hotelService = hotelService;
+        this.roomUtils = roomUtils;
+        this.locationRepository = locationRepository;
     }
 
     // get room by room id according to hotel id
@@ -34,13 +45,38 @@ public class RoomServiceImpl implements IRoomService {
     public RoomDto getRoomById(Long roomId) {
         Room room = roomRepository.findById(roomId).orElseThrow(() -> new ResourceNotFoundException("Room", "Id", roomId));
         // return dto later
-        return mapToDto(room);
+        Location location = locationRepository.findByHotelHotelId(room.getHotel().getHotelId());
+        room.getHotel().setLocation(location);
+
+        return roomUtils.mapToRoomDto(room);
     }
 
     @Override
-    public List<RoomDto> getAllRoomsByHotelId(Long hotelId) {
-        List<Room> rooms = roomRepository.findByHotelHotelId(hotelId);
-        return rooms.stream().map(this::mapToDto).toList();
+    public RoomResponse getAllRoomsByHotelId(Long hotelId, int pageNo, int pageSize, String sortBy, String sortDir) {
+
+        Sort sort = sortDir.equalsIgnoreCase(Sort.Direction.ASC.name()) ? Sort.by(sortBy).ascending() : Sort. by(sortBy).descending();
+
+        // create Pageable instance
+        Pageable pageable = PageRequest.of(pageNo, pageSize, sort);
+        Page<Room> rooms = roomRepository.findByHotelHotelId(hotelId, pageable);
+        List<Room> roomList = rooms.getContent();
+        List<RoomDto> roomDtos = roomList.stream().map(roomUtils::mapToRoomDb).toList();
+
+        return RoomResponse
+                .builder()
+                .content(roomDtos)
+                .pageNo(rooms.getNumber())
+                .pageSize(rooms.getSize())
+                .totalElements(rooms.getTotalElements())
+                .totalPages(rooms.getTotalPages())
+                .last(rooms.isLast())
+                .build();
+    }
+
+    @Override
+    public List<RoomDto> getAllRooms() {
+        List<Room> rooms =  roomRepository.findAll();
+        return rooms.stream().map(roomUtils::mapToRoomDto).collect(Collectors.toList());
     }
 
     @Override
@@ -52,14 +88,15 @@ public class RoomServiceImpl implements IRoomService {
                 .roomNumber(roomDto.getRoomNumber())
                 .description(roomDto.getDescription())
                 .capacity(roomDto.getCapacity())
-                .status(RoomType.AVAILABLE)
+                .status(roomDto.getStatus())
+                .pricePerNight(roomDto.getPricePerNight())
                 .viewType(roomDto.getViewType())
                 .hotel(hotel)
                 .build();
         // save a room
         roomRepository.save(room);
 
-        return mapToDto(room);
+        return mapToRoomDto(room);
     }
 
     @Override
@@ -74,17 +111,28 @@ public class RoomServiceImpl implements IRoomService {
         // save a room
         roomRepository.save(room);
 
-        return mapToDto(room);
+        return roomUtils.mapToRoomDto(room);
     }
 
     @Transactional
     @Override
     public void deleteRoomById(Long roomId) {
         Room room = roomRepository.findById(roomId).orElseThrow(() -> new ResourceNotFoundException("Room", "Id", roomId));
+        List<Reservation> reservationList = room.getReservationRoom().stream().map(ReservationRoom::getReservation).toList();
+        if (!reservationList.isEmpty()) {
+            throw new RuntimeException("Room is already reserved");
+        }
         roomRepository.delete(room);
     }
 
-    private RoomDto mapToDto(Room room){
+    @Override
+    public void updateStatus(Long roomId) {
+        Room room = roomRepository.findById(roomId).orElseThrow(() -> new ResourceNotFoundException("Room", "Id", roomId));
+        room.setStatus(RoomType.AVAILABLE);
+        roomRepository.save(room);
+    }
+
+    private RoomDto mapToRoomDto(Room room){
         return RoomDto.builder()
                 .roomId(room.getRoomId())
                 .roomNumber(room.getRoomNumber())
@@ -92,20 +140,9 @@ public class RoomServiceImpl implements IRoomService {
                 .capacity(room.getCapacity())
                 .status(room.getStatus())
                 .viewType(room.getViewType())
-                .build();
-    }
-
-    private RoomImageDto mapRoomImageToDto(RoomImage roomImage){
-        return RoomImageDto.builder()
-                .assetId(roomImage.getAssetId())
-                .secureUrl(roomImage.getSecureUrl())
-                .build();
-    }
-
-    private RoomImage mapRoomImageToEntity(RoomImageDto roomImageDto){
-        return RoomImage.builder()
-                .assetId(roomImageDto.getAssetId())
-                .secureUrl(roomImageDto.getSecureUrl())
+                .pricePerNight(room.getPricePerNight())
+                .reviews(room.getReviews())
+                .hotel(room.getHotel())
                 .build();
     }
